Add 4 of the 5 pedastals in the first Chamber
Boss battles at the end of Level 24, following an infodump
on Object Oriented design. I was feeling good about my progress
until these chapters knocked me on my ass again. It never
occured to me to make an array of enums for the Card. The Point
was fairly simple, as was the color, though originally I decided
to make some tuples rather than having the class instance itself.
The Locked door was a bit of fun but I fear I've overengineered it.