Add Pits expansion to Fountain of Objects
Created an alogorithm that scales the number of pits with the map
size, effectively doubling the number for each setting. The power
of interfaces started to sink in a bit for me here. Implementing
Pits to the game was relatively painless, requiring only minor
changes to the code and the creation of a new class SensePit.

The random location generator works beautifully and adding 1 or 4
pits to the game was seamless.

Removed property from SenseEmpty as empty rooms currently return
no sensory inputs and thus don't care if player is in or nearby.
Effectively a placeholder class to ensure all paths return properly.