Refactor game creation for scale
Eliminated algorithm in Grid for determining number of pits and tied
output to the game creation in Program. It means more parameters
passed into and sometimes through CreateGame but makes room for
additional start conditions without cluttering classes.

THis move paves the way for the expansion Maelstroms to be added.